Compiler warnings are cleaned up (gcc-3.2), make install works and puts all of the samples in the Common/bin/Debug directory.  The main libraries and plugins are installed in the correct paths.  Gtk+ GUI for config.

+++ PlugIns/BSPSceneManager/src/OgreBspSceneManager.cpp
@@ -515,7 +515,7 @@ namespace Ogre {
         {
             if (random)
             {
-                int idx = Math::getSingleton().UnitRandom() * mLevel->mPlayerStarts.size();
+                int idx = (int)Math::getSingleton().UnitRandom() * mLevel->mPlayerStarts.size();
                 return mLevel->mPlayerStarts[idx];
             }
             else
